Remove KeyInput, fix up project structure.
