+++ Wave Game.iml
@@ -5,13 +5,11 @@
     <exclude-output />
     <content url="file://$MODULE_DIR$">
       <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
+      <excludePattern pattern="HighScores.txt" />
     </content>
     <orderEntry type="sourceFolder" forTests="false" />
     <orderEntry type="inheritedJdk" />
-    <orderEntry type="library" name="AppleJavaExtensions" level="project" />
-    <orderEntry type="library" name="Jamepad" level="project" />
-    <orderEntry type="library" name="AppleJavaExtensions (2)" level="project" />
-    <orderEntry type="library" name="AppleJavaExtensions (3)" level="project" />
-    <orderEntry type="library" name="Jamepad (2)" level="project" />
+    <orderEntry type="library" exported="" name="AppleJavaExtensions" level="project" />
+    <orderEntry type="library" exported="" name="Jamepad" level="project" />
   </component>
 </module>
