+++ src/mainGame/GameMode.java
@@ -1,14 +1,14 @@
 package mainGame;
 
-import java.awt.Point;
-import java.awt.event.KeyListener;
-import java.awt.event.MouseAdapter;
+import java.awt.*;
+import java.awt.event.KeyEvent;
+import java.awt.event.MouseEvent;
 
 /**
  * @author Team B3
  * @author Aaron Paterson 9/12/19
  */
-public abstract class GameMode extends MouseAdapter implements Animatable {
+public abstract class GameMode extends GameState {
     private GameState state;
     public void setState(GameState s) {
         state = s;
@@ -16,8 +16,25 @@ public abstract class GameMode extends MouseAdapter implements Animatable {
     public GameState getState() {
         return state;
     }
-    abstract void resetMode();
 	abstract GameObject getEnemyFromID(ID x, Point spawnLoc);
+    abstract void resetMode();
 	abstract void resetMode(boolean hardReset);
-	abstract KeyListener getKeyInput();
+
+    @Override
+    public void keyPressed(KeyEvent e) {
+        state.keyPressed(e);
+    }
+    @Override
+    public void keyReleased(KeyEvent e) {
+        state.keyReleased(e);
+    }
+
+    @Override
+    public void mousePressed(MouseEvent e) {
+        state.mousePressed(e);
+    }
+    @Override
+    public void mouseReleased(MouseEvent e) {
+        state.mouseReleased(e);
+    }
 }
